+++ src/components/TreeRegi/Location/Map/RegistButton/index.tsx
@@ -1,0 +1,7 @@
+import Button from 'components/common/button';
+
+function RegistButton() {
+  return <Button>이 위치로 트리 등록하기</Button>;
+}
+
+export default RegistButton;
