+++ src/components/TreeRegi/Location/Search/SearchResult/style.ts
